api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Rob <robert.i.hendry@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
beautifulsoup4 = "^4.12.3"
google = "^3.0.0"
pypdf = "^4.1.0"
uvicorn = "^0.29.0"
sse-starlette = "^2.0.0"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
openai = "^1.14.3"
pyyaml = "^6.0.1"
aiostream = "^0.5.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
